Added a custom event listener on the text area for the SMTP header so that the value can be cleared when the clear button is clicked.
The overall process is to register a hook that will add an event
listener to the element and then to dispatch a matching JS event from
the button.  The downside to this approach is that it is unlikely to
play nicely with the LiveView lifecycle, but in this case I don't think
it will matter (famous last words).